update script to handle relevant non-paragraph blocks.
Prior to this commit the script only handled updating blocks of type 'paragraph'. I started with paragraph blocks because they are very common and I wanted to get a single type of block working. But now we handle all types that conceivably would contain [[...]] literals.

See the variable 'BLOCK_TYPES_TO_PROCESS' in main.py for all the block types we now update