fix bug with virtual text creation.
The virtual text creation function incorrectly assumed that the output of
re.split would result in alternating mentions and non-mentions. However
that was not the case when you had input text that looks like:
'[[blah]][[bloop]] something else'
where the two [[...]] literals were consecutive.
I used a different approach to determine which were mentions and which were not
which is more robust